requirements.txt
requests
beautifulsoup4
tqdm
pypandoc_binary # Includes pandoc
weasyprint
# pypandoc
# pandoc
joblib